refactor: Extract makeOverrides in multiple files
no code change
