+++ react/MuiCozyTheme/makeOverrides.js
@@ -1,228 +1,15 @@
-import merge from 'lodash/merge'
-
-import { alpha, lighten, darken } from '../styles'
+import { alpha, darken } from '../styles'
+import {
+  makeAlertColor,
+  makeChipStyleByColor,
+  makeSecondaryButtonStyle,
+  makeTextButtonStyle,
+  makeContainedButtonStyle
+} from './helpers'
 
 const SWITCH_BAR_WIDTH = 25
 
-export const makeThemeOverrides = theme => {
-  const createOverrides =
-    theme.palette.type === 'dark' ? makeInvertedOverrides : makeOverrides
-
-  return createOverrides(theme)
-}
-
-const makeAlertColor = (theme, color) => {
-  const themeColorByColor = {
-    primary: theme.palette[color].main,
-    secondary: theme.palette.text.primary
-  }
-
-  // same approach as Mui, see https://github.com/mui/material-ui/blob/v4.x/packages/material-ui-lab/src/Alert/Alert.js#L28
-  return {
-    '&-standard': {
-      color: darken(themeColorByColor[color], 0.6),
-      backgroundColor: lighten(themeColorByColor[color], 0.9),
-      '& $icon': {
-        color: themeColorByColor[color]
-      },
-      '& $action': {
-        '& button[title="Close"]': {
-          color: theme.palette.text.secondary
-        }
-      }
-    },
-    '&-outlined': {
-      color: darken(themeColorByColor[color], 0.6),
-      border: `1px solid ${themeColorByColor[color]}`,
-      '& $icon': {
-        color: themeColorByColor[color]
-      }
-    },
-    '&-filled': {
-      backgroundColor:
-        color === 'secondary'
-          ? theme.palette.grey[600]
-          : themeColorByColor[color]
-    }
-  }
-}
-
-const makeAlertInvertedColor = (theme, color) => {
-  return {
-    '&-standard': {
-      color: theme.palette.primary.main,
-      backgroundColor: theme.palette.background.default,
-      '& $icon': {
-        color: theme.palette[color].main
-      }
-    },
-    '&-outlined': {
-      color: theme.palette.primary.main,
-      border: `1px solid ${theme.palette.primary.main}`,
-      '& $icon': {
-        color: theme.palette[color].main
-      }
-    },
-    '&-filled': {
-      color: theme.palette[color].contrastText,
-      backgroundColor:
-        color === 'secondary'
-          ? theme.palette.grey[200]
-          : theme.palette[color].main,
-      '& $icon': {
-        color: theme.palette[color].contrastText
-      }
-    }
-  }
-}
-
-const makeChipStyleByColor = (theme, color) => ({
-  color: theme.palette.text[color] || theme.palette[color].main,
-  borderColor:
-    color === 'primary'
-      ? theme.palette.border.main
-      : alpha(theme.palette[color].main, theme.palette.border.opacity),
-  '&$clickable, &$deletable': {
-    '&:hover, &:focus': {
-      borderColor:
-        color === 'primary'
-          ? theme.palette.border.main
-          : alpha(theme.palette[color].main, theme.palette.border.opacity),
-      backgroundColor:
-        color === 'primary'
-          ? theme.palette.action.hover
-          : alpha(theme.palette[color].main, theme.palette.action.hoverOpacity)
-    }
-  },
-  '& $icon': {
-    color:
-      color === 'primary' ? theme.palette.text.icon : theme.palette[color].main,
-    fill:
-      color === 'primary' ? theme.palette.text.icon : theme.palette[color].main
-  },
-  '& $deleteIcon': {
-    color:
-      color === 'primary'
-        ? theme.palette.text.secondary
-        : theme.palette[color].main,
-    fill:
-      color === 'primary'
-        ? theme.palette.text.secondary
-        : theme.palette[color].main
-  },
-  '&$colorPrimary': {
-    padding: '0 1px',
-    color: theme.palette[color].contrastText,
-    backgroundColor: theme.palette[color].main,
-    '& $icon, & $deleteIcon': {
-      color: theme.palette[color].contrastText,
-      fill: theme.palette[color].contrastText
-    },
-    '&$disabled': {
-      opacity: 1,
-      color: theme.palette.text.disabled,
-      backgroundColor: theme.palette.action.disabledBackground,
-      '& $icon, & $deleteIcon': {
-        color: theme.palette.text.disabled,
-        fill: theme.palette.text.disabled
-      }
-    },
-    '&$clickable, &$deletable': {
-      '&:hover, &:focus': {
-        backgroundColor: theme.palette[color].dark
-      }
-    }
-  },
-  '&.ghost': {
-    borderWidth: '1px',
-    borderStyle: 'dashed',
-    '&:not($disabled)': {
-      color: theme.palette[color].main,
-      borderColor: alpha(
-        theme.palette[color].main,
-        theme.palette.border.ghostOpacity
-      ),
-      backgroundColor: alpha(
-        theme.palette[color].main,
-        theme.palette.action.ghostOpacity
-      ),
-      '& $icon, & $deleteIcon': {
-        color: theme.palette[color].main,
-        fill: theme.palette[color].main
-      }
-    },
-    '&$clickable, &$deletable': {
-      '&:hover, &:focus': {
-        borderColor: alpha(
-          theme.palette[color].main,
-          theme.palette.border.ghostOpacity
-        ),
-        backgroundColor: alpha(
-          theme.palette[color].main,
-          theme.palette.action.hoverGhostOpacity
-        )
-      }
-    }
-  }
-})
-
-const makeSecondaryButtonStyle = (theme, color) => ({
-  color: theme.palette[color].main,
-  borderColor: theme.palette[color].main,
-  '&:hover': {
-    backgroundColor: alpha(
-      theme.palette[color].main,
-      theme.palette.action.hoverOpacity
-    ),
-    '@media (hover: none)': {
-      backgroundColor: 'transparent'
-    }
-  },
-  '&.ghost': {
-    backgroundColor: alpha(
-      theme.palette[color].main,
-      theme.palette.action.ghostOpacity
-    ),
-    '&:hover': {
-      backgroundColor: alpha(
-        theme.palette[color].main,
-        theme.palette.action.hoverGhostOpacity
-      ),
-      '@media (hover: none)': {
-        backgroundColor: alpha(
-          theme.palette[color].main,
-          theme.palette.action.ghostOpacity
-        )
-      }
-    }
-  }
-})
-
-const makeTextButtonStyle = (theme, color) => ({
-  color: theme.palette[color].main,
-  '&:hover': {
-    backgroundColor: alpha(
-      theme.palette[color].main,
-      theme.palette.action.hoverOpacity
-    ),
-    '@media (hover: none)': {
-      backgroundColor: 'transparent'
-    }
-  }
-})
-
-const makeContainedButtonStyle = (theme, color) => ({
-  color: theme.palette[color].contrastText,
-  backgroundColor: theme.palette[color].main,
-  '&:hover': {
-    backgroundColor: theme.palette[color].dark,
-    '@media (hover: none)': {
-      backgroundColor: theme.palette[color].main
-    }
-  }
-})
-
-const makeOverrides = theme => ({
+export const makeOverrides = theme => ({
   MuiOutlinedInput: {
     root: {
       borderRadius: 4,
@@ -1013,111 +800,3 @@ const makeOverrides = theme => ({
     }
   }
 })
-
-const makeInvertedOverrides = invertedTheme => {
-  const makeOverridesForInvertedTheme = invertedTheme => ({
-    MuiOutlinedInput: {
-      root: {
-        boxSizing: 'border-box',
-        '&$disabled': {
-          background: 'initial'
-        },
-        '&$focused $notchedOutline': {
-          borderColor: invertedTheme.palette.text.primary,
-          borderWidth: '0.0625rem'
-        },
-        '& $notchedOutline': {
-          borderColor: invertedTheme.palette.text.primary
-        }
-      }
-    },
-    MuiFormLabel: {
-      root: {
-        '&$focused': {
-          color: invertedTheme.palette.text.primary
-        }
-      }
-    },
-    MuiLinearProgress: {
-      colorPrimary: {
-        backgroundColor: 'rgba(255,255,255,0.2)'
-      },
-      colorSecondary: {
-        backgroundColor: 'rgba(255,255,255,0.2)'
-      }
-    },
-    MuiCheckbox: {
-      colorPrimary: {
-        '&$checked:not($disabled)': {
-          color: invertedTheme.palette.primary.light
-        }
-      },
-      colorSecondary: {
-        '&$checked:not($disabled)': {
-          color: invertedTheme.palette.error.main
-        }
-      }
-    },
-    MuiAlert: {
-      root: {
-        '&.cozyAlert': {
-          '&-primary': makeAlertInvertedColor(invertedTheme, 'primary'),
-          '&-secondary': makeAlertInvertedColor(invertedTheme, 'secondary'),
-          '&-success': makeAlertInvertedColor(invertedTheme, 'success'),
-          '&-error': makeAlertInvertedColor(invertedTheme, 'error'),
-          '&-warning': makeAlertInvertedColor(invertedTheme, 'warning'),
-          '&-info': makeAlertInvertedColor(invertedTheme, 'info')
-        }
-      }
-    },
-    MuiSnackbarContent: {
-      root: {
-        backgroundColor: invertedTheme.palette.grey[200]
-      }
-    },
-    MuiTabs: {
-      root: {
-        '&.segmented': {
-          '& $indicator': {
-            backgroundColor: invertedTheme.palette.primary.main
-          }
-        }
-      }
-    },
-    MuiTab: {
-      root: {
-        '&.segmented': {
-          '&$selected': {
-            color: invertedTheme.palette.primary.contrastText
-          }
-        }
-      }
-    },
-    MuiSwitch: {
-      switchBase: {
-        color: invertedTheme.palette.grey[100]
-      },
-      colorPrimary: {
-        '&$checked': {
-          '& + $track': {
-            backgroundColor: invertedTheme.palette.success.dark
-          }
-        }
-      },
-      colorSecondary: {
-        '&$checked': {
-          '& + $track': {
-            backgroundColor: invertedTheme.palette.success.dark
-          }
-        }
-      }
-    }
-  })
-
-  const invertedOverrides = merge(
-    makeOverrides(invertedTheme),
-    makeOverridesForInvertedTheme(invertedTheme)
-  )
-
-  return invertedOverrides
-}
